Updated project to support message sending functionality
- Updated Dockerfile to use Go 1.20, ensuring compatibility with project
dependencies.
- Modified index.html to include login/logout buttons and a container for
sending messages. This enhances the user interface by allowing users to
interact with Twitch and YouTube services more seamlessly.
- Updated main.js with several important changes:
  - The checkSession function now accurately reflects the user's login status,
improving reliability in session management.
  - The updateUIForLoggedInUser and updateUIForLoggedOutUser functions have
been revised to dynamically adjust the UI based on the user's authentication
status with Twitch and YouTube, ensuring a more intuitive user experience.
  - Added functionality for Twitch and YouTube login buttons, streamlining
the login process and making the application more user-friendly.
  - Implemented the sendMessage function, enabling users to send messages to
Twitch and YouTube chats directly from the web interface, significantly
enhancing interactivity.
- Made significant revisions to auth.go:
  - The callbackHandler now properly handles OAuth callbacks from Twitch and
YouTube, improving the robustness of the login flow.
  - Updated the sessionMiddleware to ensure that requests are properly
authenticated, enhancing security and session management.
  - Introduced send.go, a new file that contains the backend logic for sending
messages to Twitch and YouTube. This addition is a crucial component of the
message-sending feature, allowing users to engage with their audience in
real-time.

These changes collectively enhance the application's functionality by providing
a seamless experience for logging in, managing sessions, and interacting with
Twitch and YouTube chats. The inclusion of a message sending feature marks a
significant milestone in the project's development, offering streamers a more
integrated and interactive platform for communication with their viewers.